Save information about opened/closed FILE streams into the trace. This is the first step of issue #37.
Add two new events: FileOpen and FileClose. These both contain the runtime
value of the FILE * for the stream, the offset of the null terminated filename
in the trace's data, and the offset of the null terminated mode string in the
trace's data. To support this, TraceStream has been updated to hold the offset
of the filename and mode strings. Two convenience functions have been added
to TraceThreadListener to handle the recording of stream opening and closing,
recordStreamOpen() and recordStreamClose(), and all listening points that
update the stream information have been updated to use these.

TraceProcessListener now records the initial runtime values for stdin, stdout,
and stderr, and records them into the process trace. These are restored by
ProcessTrace.

The trace format version has been bumped to 3.