Render FILE * values correctly. Close #37.
Add a new subclass of seec::cm::Value, seec::cm::ValueOfPointerToFILE. This is
used to represent values with type FILE *. It doesn't allow clients to access
dereferences of the pointer, or get information about the pointee, but it does
allow the client to check if the FILE * references a currently valid (open)
stream.

Update graph layout and expansion to correctly deal with ValueOfPointerToFILE.
Currently the string value is displayed, rather than rendering pointers to a
common FILE node.