Allow InputBufferAllocator to open traces directly from a trace archive.
The decision is based on the path: if it ends in ".seec" or ".seecrecord", then it is assumed to be a trace archive. Otherwise, it will be opened as a trace directory.

The implementation follows from seec-view's OpenTrace's implementation: contents are extracted to a temporary directory, then deleted when the InputBufferAllocator is destroyed.