Only allow a single instance of the trace viewer (per user) at any one time.
This is to prepare for transferring action recordings to the server in the background (we don't want multiple instances transferring the same recordings).

If a new instance attempts to open, instead forward the request to the existing instance (either opening the requested files, or showing the welcome window if no files were requested).