Add support for attributes with Decl * or Stmt * values.
To serialize these values correctly we need access to the mapping information. This is provided by passing a reference to the seec::cm::ProcessTrace to IAttributeReadOnly::to_string_impl() and IAttributeReadWrite::from_string_impl(). The values are written as a string that is intended to be used for both deserializing and for human readability: "[MappedAST index] [Node index] [Filename] [Line]:[Column]".

The deserialization requires EventHandler to have access to the mapping information (again exposed via the seec::cm::ProcessTrace). ActionReplayFrame now holds a reference to the seec::cm::ProcessTrace, which allows the interface to ActionReplayFrame::RegisterHandler() to remain unchanged.