Add seec::make_function(), which wraps callables into a std::function by automatically deducing the appropriate type.
Callable objects (including lambdas) have the appropriate type deduced from the operator() member. Note that this will only work if the type is not ambiguous.

seec::make_function() can also be used to conveniently wrap a class instance with a pointer to one of the class's member functions, into a std::function which simply calls the member on the given class instance.