Add a new method for formatting Stmt text that uses tokens from the original source code rather than pretty-printing the Stmt.
This system uses does not expand macros if they were defined in system headers, in order to hide the messy details from students. It does expand macros that weren't defined in system headers, so that students can see the full behaviour of their own macros.

This system currently only works for Stmts that are in the main source file. If the Stmt locations can't be found, are in a non-main source file, or start and end in different source files, then we fall back to the old pretty-printing system.