Update formatStmtSource() method so that it handles Stmts that are not written in the main file (i.e. functions that are defined in an #included file).
Also prepare for caching tokens outside of the formatter, so that we can avoid repeatedly preprocessing / raw lexing the source files. At the moment the per-AST cache is only used internally - the next step will be to expose a cache manager that can be passed in to the formatter, e.g. to be held by StateEvaluationTree.