Remove the highlight notification for values produced by the active Stmt from the SourceViewer.
Add highlights for values produced by the active Stmt to the StateGraphViewer. Because the highlighting is performed by the graph viewer itself, these now use the same colour scheme as other "active value" displays (i.e. the source viewer and the dynamic evaluation tree).

Also update the graph viewer's highlighting for pointer values to also highlight the edge and the direct pointee (this allows us to see the pointee highlighted when moving the mouse over an rvalue pointer in the dynamic evaluation tree).

Also update the SourceViewer's determination of ranges to use getExpansionRange() so that function-like macros are highlighted correctly, and to avoid unnecessary loops when not getting the immediate expansion range.