Several updates to seec::Maybe:
Use ref-qualifiers so that calling get() on an rvalue Maybe returns an rvalue reference to the accessed element.

Use a template alias for MaybeValue, ValueAt, to clean up some of Maybe's code.

Use a constexpr template helper for MaybeIndexByType, IndexOf, to clean up some of Maybe's code.

Only allow get() to set the active element if the element is default constructible, and run the default constructor in this case.