Correct pointer object tracking through qsort() and bsearch().
Modify TracedFunction to allow the creation of "shim" objects, which have no FunctionIndex. These are used to represent qsort() and bsearch() themselves. A shim TracedFunction's pointer object information for llvm::Arguments holds the pointer objects for the arguments that were passed to the shim's called function (i.e. for qsort() the shim holds the pointer object information for the pointers received by the comparison function).

Also fix a bug with bsearch() where the result was not being returned, or associated with the call Instruction.