Some work towards supporting arrays with variably sized element types.
Add the option to pass a seec::trace::FunctionState to getValue() for in-memory values, defining the function that "owns" the value (e.g. the function that a local or parameter is contained in).

Store a reference to the seec_clang::MappedModule on ValueStoreImpl, so that it can be used within associated Values.

Attempt to caculate the size of variable length array element types when creating new in-memory array Values (ValueByMemoryForArray), by getting the runtime value of each VariableArrayType's size Expr.