When checking pointer validity, only use the MemoryArea of the allocation that the pointer is referencing (i.e. that which contains the pointer object's base).
If this MemoryArea does not exist, then raise a PointerObjectOutdated error. This improves the accuracy of error messages when using pointers that refer to deallocated memory (when the area has not been reallocated to a new allocation), so they accurately state that the pointer's target object has been deallocated, rather than stating that the pointer no longer refers to its target object.