Blatantly use internal ICU method.
Use icu::ResourceBundle::getWithFallback() rather than regular get(), which allows us to fall back to a less-precise locale when attempting to access nested resources. The ICU documentation indicates that this should not be used, but there is no public alternative available at the moment. If it becomes a problem later on then we can implement our own solution.