Add replay for mouseover value in graph.
This require a method to find a Value given an address and a string representation of the type. However, we know that the Value must exist in the current state's ValueStore in order for the user to mouseover its representation in the graph window, so this is achieved by searching the current state's ValueStore using the newly added method findFromAddressAndType.