Add augmentation for ClangEPV messages.
This is the start of the system that will split out institution-specific
content so that SeeC is easier to customize and redeploy. It replaces
references to lectures with a marker called an "augmentation" such as:
  $[concept:struct]

In this example the augmentation's name/type is "concept", and its
identifier is "struct". The explanation system can be provided a
callback function which accepts a name and identifier, and returns
a string which should replace the augmentation marker. If a client
does not provide such a callback, then all augmentation markers are
simply removed from the explanation. The augmentation process occurs
before formatting and indexing, so augmentation text can include needles
that link to lecture notes etc.

The trace viewer's AugmentationCollection reads augmentation information
defined in XML files. These files are structured as so:
<?xml version="1.0" encoding="UTF-8"?>
<package name="Foo">
  <augmentations locale="root">
    <concept id="struct">Read more about structures at...</concept>
  </augmentations>
</package>

Currently these files are read from the "augment" directory in SeeC's
resources directory, e.g. "/usr/local/lib/seec/resources/augment/*.xml"

The augmentation text from all files is concatenated, with a newline
character separating each one. This is currently used in the explanation
viewer and the Stmt tooltip.