Use common wxConfig between applications.
seec-print and seec-view now use a common data directory for user
information, including configuration. This will allow the user to select
settings in the trace viewer, and have them honoured in the other
applications (and soon in the runtime library).

This commit also enables augmentation for runtime error messages in
seec-print, loading augmentations from the resource directory.