Load augmentations from the user's data directory.
Augmentations are loaded from the folder "augment" in the user's local
data directory.

Previously setupCommonConfig() used a temporary wxAppConsole (if there
wasn't already a running wxApp) which was destructed at the end of the
function. It's now necessary for clients to setup a wxApp instance. The
function "setupDummyAppConsole()" is now public to facilitate this for
clients that don't care about the particular instance (e.g. seec-print
and the runtime library), and now automatically registers the
appropriate shutdown function using std::atexit().