Display a table of active augmentations in the preferences dialog.
Allow registering listeners of an AugmentationCollection to be notified when
the collection changes (currently only when a new Augmentation is added).

Add AugmentationCollectionDataViewModel, which implements a
wxDataViewVirtualListModel (and uses the aforementioned listener interface).

Add a new library, SeeCwxWidgetsGUI, for holding wxWidgets related classes which
will not be used from command line programs (or the runtime library). Currently
this is only used for AugmentationCollectionDataViewModel, but it prevents
seec-print and the runtime library from depending on wxWidgets' core and adv.