Use only the most recent version of augmentations.
If there are multiple augmentation documents with the same ID, then only
use the document with the highest version when performing augmentation.
The unused augmentations are indicated by greyed (disabled) text in the
augmentations preference panel, but can still be interacted with
(deleted). If an active augmentation is deleted then it's automatically
replaced by the most recent version (if any). Similarly, if a newer
version of an augmentation is downloaded it automatically replaces the
existing version.