Allow user to enable/disable augmentations.
This is stored in the user's configuration on a per-ID per-version
basis, so disabling one instance of an (ID,version) document will
disable all instances of it. Any differences in an augmentation document
should be indicated by an updated version number anyway, so I don't
foresee this being a problem.

If an active augmentation is disabled, the next highest enabled version
of that augmentation will become active (if it exists). Similarly, if an
augmentation is enabled and it has a higher version than the currently
active augmentation for its ID, then it will immediately become active.