Improve GetCurrentRuntimeValueAsImpl<long double>.
Correct implementation for Instructions to retrieve the long double value of the
runtime value. Correct implementation for ConstantFPs to handle IEEEsingle,
IEEEdouble, and x87DoubleExtended. The latter should only be used if the host's
long double uses this format (our rough assertion at the moment is checking
that LDBL_MANT_DIG == 64).