Start removing use of getCurrentRuntimeValueAs.
Fix GetMemoryOfBuiltinAsString<long double> to work, but only for x87 80 bit
long doubles at the moment. (Previously the reported size of long double, 16
bytes, would cause the value to be reported as uninitialized - even if the 10
bytes that are actually used were initialized). This also requires a manual
adjustment of the size before constructing the ValueByMemoryForScalar (we simply
check if the type's float semantics are x87DoubleExtended, and if so we set the
size to be 10 bytes).

Update GetValueOfBuiltinAsString to use GetRecreatedValue.hpp methods getAPInt,
getAPSInt, and getAPFloat.