Don't use uintptr_t for recreated pointers.
This is one part of making execution traces portable across architectures and
systems. For traced programs, we should only use uintptr_t at runtime. We use
the type seec::trace::stateptr_ty to represent "recreated" pointers, which is
currently defined to be uint64_t. This type must be big enough to represent
any recreated pointer value (eventually we should ensure that the trace itself
can represent any value, but uint64_t should be fine for the recreated state
for now). This patch updates all uses of uintptr_t in the recreated state,
Clang-mapped state, and seec-trace-view to use stateptr_ty. This type is also
redeclared as seec::cm::stateptr_ty for convenience.

The MemoryArea class in DSA is unfortunately shared between the shadow state
and the recreated state. For now we are defining MemoryArea to use uint64_t
rather than uintptr_t, but in the future we should use separate classes to
represent memory areas in the shadow state and recreated state.

Update ValueByMemoryForPointer to read the value from memory based on the size
of the pointer type in the ASTContext, rather than the size of a pointer on the
host platform. There are a few other places in MappedValue.cpp where platform
dependent types are used - these will be corrected soon.

Also remove the MemoryBlock class, which was no longer used.