Add testing for comparable trace states.
This adds tests for the expected comparable states for all execution traces in
the test suite. Any trace that is printed with seec_test_print_trace() is
expected to have a corresponding ".expected" file which contains the expected
output of printing the trace with "seec-print -S -comparable -reverse". This
should help us to maintain order when making changes to the execution tracing
system in the future!