Remove PreviousSame field from instruction events.
When rewinding instruction events, find the preceding event by searching the
execution trace for it. This is significantly slower, but in the future we'll
store the previous values in the state so that rewinding is faster and to
remove the need to store all values in the trace (similarly to how memory states
are stored on the state now). Removing this field significantly reduces the
size of many execution traces (combined with the previous removal of the
thread time field, up to 66% size reduction for uncompressed traces in tests).