Always enable action recording.
This is to allow the creation of custom action recordings, which instructors or
students can distribute manually. The SEEC_USER_ACTION_RECORDING definition is
still used to control whether or not recordings are automatically submitted via
HTTP to seeclogd (still hardcoded to the CSSE server, but in the future we'll
open this up for customization).