Add hover tooltips to the source viewer.
This allows users to hover over a Decl or Stmt in the source code to receive an
explanation of the node. The Stmt explanation doesn't include any information
related to the active function's state, so it is not as informative as the
tooltip produces by hovering over the dynamic evaluation tree.