Allow trace archives to contain an "annotations.xml" file.
This is recognised by OpenTrace and loaded as an AnnotationCollection. The
AnnotationCollection class is simply a thin wrapper around the underlying
wxXmlDocument, and currently does nothing. In the future this will be used to
allow annotating execution traces with additional information for particular
AST nodes, thread states, and process states.