Support indexing in ProcessState and ThreadState annotations.
This uses seec::icu::IndexedString internally, so the syntax of indexing a
string is exactly the same as for ClangEPV. The difference is that we don't
support relative node names, because these strings refer to an entire state and
not a specific AST node, so we have an absolute node referencing scheme:
  decl:ASTIndex,DeclIndex
  stmt:ASTIndex,StmtIndex

For example we might use the index "@[stmt:0,16]this call@[]".

The class IndexedAnnotationText is used to handle the initial indexing and
searching for the primary index at a given character position. The search result
is an AnnotationIndex object, which handles the lookup of Decl and Stmt nodes
from the syntax above.