Don't suffix strings with elision cells.
Change when elision cells are generated so that they're only used to
separate value cells (they aren't placed if elision lasts until the end
of the array). Also, don't generate a port for elision cells because we
no longer support custom ports (and the elided characters are, by
definition, not referenced).