Extend contextual navigation for pointers.
Add the option to rewind to the most recent point at which a pointer's
referenced memory was allocated (including file and directory stream
pointers). For example, if a runtime error is raised for a double-free,
this can be used to rewind to the point at which the free'd pointer was
previously valid (i.e. immediately prior to the first free).