Enable seec-cc, seec-ld, and runtime for windows.
Record info is fundamentally different, because on windows we can't
declare symbols in the runtime library that will be linked to symbols in
the executable. However, the changes are minimal:
- only declare symbols in RecordInfo.h under __unix__ or __APPLE__.
- when the globals are added to the executable (in RecordExternal.cpp),
set their DLLStorageClass to DLLExportStorageClass to ensure that they
will be visible from the runtime library.
- in the runtime library, lookup the symbols dynamically using
GetProcAddress().

The remaining changes to enable compilation and linking are
straightforward: use seec-ld.exe instead of seec-ld, find the runtime
library in the same directory as seec-cc (as opposed to ../lib/seec),
etc.