Correct graph display for windows.
The environment variables we generate for calling dot under Linux and OS
X are invalid for windows, but aren't needed either, so simply don't
pass an environment block.

Also, ensure that the IE emulation setting for seec-view.exe is the
highest version possible. The default for applications using WebBrowser
is IE7, which doesn't handle the graph SVG.