Allow user-configurable colour schemes.
Currently the colour scheme is only used by the displays which use
wxStyledTextCtrl, i.e. the source code, the file stream output, and the
explanation viewer. The user can configure the colour scheme through the
preferences menu. All changes are detected and applied by the displays
immediately, allowing the user to observe the results. If the user exits the
preference menu without saving, then the prior colour scheme is restored. If
the preference menu is exited with saving, then the new colour scheme is saved
into an XML file in the user's local data directory, which will be loaded and
used the next time the trace viewer is opened.