Only convert header search path to native.
Based on commit messages, this was supposed to be in
19a6675378f853bbdc80a6d85bdb1671a1ee27cc and somehow got lost.
Reproduced commit message is below:

When converting filenames to the native format, only convert the section
of a filename which matches a header search path. This
1) prevents us converting filenames which aren't converted elsewhere
(e.g. the name of the main source file if it uses an absolute path).
2) avoids an issue where relative paths aren't found under msw if the
file name is completely converted to the native format (e.g. #include
<sys/something.h>).