Make preferences look nicer under MSW.
Making the FG and BG colour pickers expand to match the height of the
font selector makes all buttons uniform height, which looks nicer.

The wxBitmapComboBox in the LocaleSettings was getting squished
vertically under MSW. Removing the border from the top and bottom seems
to correct this effectively for now. Under other platforms, add a spacer
above the combo to achieve the same effect that the border had.