Hide dot.exe execution on Win32.
This feature isn't exposed through LLVM's general function for executing
external programs. It is by the wxWidgets interface, but that can only be used
from the main thread, and we execute dot in a worker thread. For now we will use
a slightly adapted version of the LLVM function's Win32 implementation for Win32
builds, and the general interface for all other platforms.