Make shadow memory allocation-based rather than fragment-based.
This is in preparation for bit-level initialization tracking
(mheinsen issue #9). This also simplifies the tracking of
initialization within allocations, though we do have to ensure
that allocations are added and removed correctly (and we never
try to set or clear the initialization of memory if the
allocation does not exist).

Note that this brings the tracing system's handling into line
with the replay system, which became allocation-based back in
commit 97b8557fca6edaaa967b0528e3a69ca36774c8a8 !