Change recreated runtime value storage.
In the new system, runtime values are stored separately for each BasicBlock in
a Function. When we jump backwards, the cleared BasicBlocks are kept in storage,
then when we rewind the backwards jump they can be pulled out of storage and
restored. Thus we don't need to search for the previous value of an Instruction
when we rewind an Instruction event.

This makes rewinding the trace a lot faster. However, the real purpose of this
change is to prepare for the ability to set runtime values without any
information from the trace files (e.g. by re-calculating values based on an
Instruction and the current runtime values of its operands).