Add InstrIndexInFn, an opaque typdef used to represent the index of an Instruction in a Function.
This uses foonathan's type_safe library, which is also added as a submodule in
this commit. See more about the library here:
  https://github.com/foonathan/type_safe

Several areas of the code have been updated to use this type instead of uint32_t
to represent Instruction indices. However, it may not cover all representations.
This was an initial step to try out type_safe in SeeC's code base. As it seems
promising, we will probably use more of type_safe's features moving forward.

This commit also changes several uses of seec::Maybe<> with a single element
into uses of llvm::Optional<>. For single elements, Optional is somewhat easier
to work with (using explicit conversion to bool, and the dereferencing
operators).

Finally, a few changes are simply correcting lines that were over 80 columns.