Correct assertion on re-setting a runtime value in a BasicBlockStore.
It's not an error to set a value twice, as it will happen if we rewind a value
being set, and then forward over the value being set again. The rewind doesn't
"unset" the value in the BasicBlockStore (there is no need to). So instead of
asserting that we are not setting a value multiple times on the same
BasicBlockStore, instead assert that we always set the *same* value.