Add a basic observer pattern template, and use it to implement ColourSchemeSettings' notifications.
Previously, listeners on ColourSchemeSettings would not correctly deregister
themselves when their lifetime ended (actually there was no way for them to
deregister), which would cause bad things to happen if the user edited the
colour scheme after closing any trace windows.

The new observer utility should be usable in several places throughout the GUI
to avoid rewriting this kind of logic (and avoid such problems in future). For
the moment it's a very basic implementation, not particularly efficient and does
not support use from multiple threads.