Add options to compile, run to the source editor.
This is still work in progress. Notably, the option to run is designed only for
linux (using gnome-terminal to do the execution so that the user can interact
with their program). In the future we'll allow linux users to select which
terminal emulator they want to view their programs in, and we'll also support
Windows (should be relatively soon) and OS X (may have to wait until I have
access to an OS X development environment).

The compile option is also quite limited at the moment, as it always attempts
to compile a single source file directly to a program. At some point in the
future it would be nice to support multi-file projects.

Of course, we'll also need to add standard options such as customising the
compiler flags that should be used, and passing arbitrary arguments to the
user's program when it is run externally.