Correct external compilation under MSW.
The most troublesome part of this is setting the path correctly so that the
Clang driver logic will detect the MinGW toolchain. We add a Compiler
preferences section to the global preferences menu, which allows the user to
pick the location of gcc.exe in their MinGW installation, then we use this to
determine the correct directory to add to the PATH of the external compiler
process. However, there were issues when attempting to copy the entire
environment from the current process (seec-view) and then modify only the PATH
for the compiler process, so instead we start with a new environment and just
copy over a selection of environment variables, and create a completely new
PATH which only has seec's bin and MinGW's bin.