Correct Maybe for newer versions of clang.
When assigning a non-reference const type which matched the non-const head type, newer versions of clang would select the more-general template member function assign, redirecting to the next element of the Maybe (leading eventually to failure). This patch disables that overload of assign when the assigning type is a const version of the Head type, thus forcing the assign to be redirected to the Head object.