Record catchable signals into the trace.
On unix platforms, install a signal handler for catchable signals which have a
default behaviour that would terminate the process. The signal handler will
record information about the signal (number, name, strsignal() result), as well
as the thread it was caught in (and the current thread time of that thread).
The signal handler then restores the default behaviour and raises the signal,
thus causing the process to terminate as expected.

This closes #14.

Don't acquire a memory lock prior to calling pthread_join, because this will
obviously cause a deadlock if the target thread needs to access memory before
completing.